requirements.txt
flask
pandas
celery
python-dotenv
redis